SAK-12944
git-svn-id: https://source.sakaiproject.org/svn/sam/trunk@48757 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/facade/AssessmentFacadeQueries.java
@@ -1435,7 +1435,7 @@ public class AssessmentFacadeQueries extends HibernateDaoSupport implements
 
 	private String fileSizeInKB(int fileSize) {
 		String fileSizeString = "1";
-		int size = Math.round(fileSize / 1024);
+		int size = Math.round((float)fileSize / 1024.0f);
 		if (size > 0) {
 			fileSizeString = size + "";
 		}
